mobile_application/cloud-proxy/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
google-adk>=0.3.0
